refactor(cli): group inference helpers
