+++ scripts/check-legacy-migrated-paths.ts
@@ -27,9 +27,9 @@ const REMOVED_SHIM_MOVES: Record<string, string> = {
   "bin/lib/chat-filter.js": "src/lib/chat-filter.ts",
   "bin/lib/config-io.js": "src/lib/state/config-io.ts",
   "bin/lib/debug.js": "src/lib/diagnostics/debug.ts",
-  "bin/lib/inference-config.js": "src/lib/inference-config.ts",
-  "bin/lib/local-inference.js": "src/lib/local-inference.ts",
-  "bin/lib/nim.js": "src/lib/nim.ts",
+  "bin/lib/inference-config.js": "src/lib/inference/config.ts",
+  "bin/lib/local-inference.js": "src/lib/inference/local.ts",
+  "bin/lib/nim.js": "src/lib/inference/nim.ts",
   "bin/lib/onboard-session.js": "src/lib/onboard-session.ts",
   "bin/lib/platform.js": "src/lib/platform.ts",
   "bin/lib/preflight.js": "src/lib/preflight.ts",
