+++ scripts/ts-migration-assist.ts
@@ -22,13 +22,13 @@ const SPECIAL_REWRITES: Record<string, Array<[string, string]>> = {
   "bin/lib/onboard.js": [
     ['require("./runner")', 'require("../../bin/lib/runner")'],
     ['require("./sandbox-build-context")', 'require("../../bin/lib/sandbox-build-context")'],
-    ['require("./local-inference")', 'require("../../bin/lib/local-inference")'],
-    ['require("./inference-config")', 'require("../../bin/lib/inference-config")'],
+    ['require("./inference/local")', 'require("../../bin/lib/local-inference")'],
+    ['require("./inference/config")', 'require("../../bin/lib/inference-config")'],
     ['require("./platform")', 'require("../../bin/lib/platform")'],
     ['require("./resolve-openshell")', 'require("../../bin/lib/resolve-openshell")'],
     ['require("./credentials/store")', 'require("../../bin/lib/credentials")'],
     ['require("./state/registry")', 'require("../../bin/lib/registry")'],
-    ['require("./nim")', 'require("../../bin/lib/nim")'],
+    ['require("./inference/nim")', 'require("../../bin/lib/nim")'],
     ['require("./onboard-session")', 'require("../../bin/lib/onboard-session")'],
     ['require("./policies")', 'require("../../bin/lib/policies")'],
     ['require("./usage-notice")', 'require("../../bin/lib/usage-notice")'],
@@ -42,7 +42,7 @@ const SPECIAL_REWRITES: Record<string, Array<[string, string]>> = {
     ['require("./lib/registry")', 'require("../bin/lib/registry")'],
     ['require("./lib/nim")', 'require("../bin/lib/nim")'],
     ['require("./lib/policies")', 'require("../bin/lib/policies")'],
-    ['require("./lib/inference-config")', 'require("../bin/lib/inference-config")'],
+    ['require("./lib/inference/config")', 'require("../bin/lib/inference-config")'],
     ['require("./lib/version")', 'require("../bin/lib/version")'],
     ['require("./lib/onboard-session")', 'require("../bin/lib/onboard-session")'],
     ['require("./lib/runtime-recovery")', 'require("../bin/lib/runtime-recovery")'],
