+++ src/lib/actions/sandbox/doctor.ts
@@ -10,8 +10,8 @@ import path from "node:path";
 import { CLI_DISPLAY_NAME, CLI_NAME } from "../../cli/branding";
 import { isErrnoException } from "../../core/errno";
 import { recoverNamedGatewayRuntime } from "../../gateway-runtime-action";
-import { probeProviderHealth } from "../../inference-health";
-import { parseGatewayInference } from "../../inference-config";
+import { probeProviderHealth } from "../../inference/health";
+import { parseGatewayInference } from "../../inference/config";
 import { stripAnsi } from "../../adapters/openshell/client";
 import { captureOpenshell } from "../../adapters/openshell/runtime";
 import { OPENSHELL_PROBE_TIMEOUT_MS } from "../../adapters/openshell/timeouts";
