+++ src/lib/actions/sandbox/status.ts
@@ -6,9 +6,9 @@
 import { CLI_DISPLAY_NAME, CLI_NAME } from "../../cli/branding";
 import { parseSandboxPhase } from "../../state/gateway";
 import { getNamedGatewayLifecycleState } from "../../gateway-runtime-action";
-import { parseGatewayInference } from "../../inference-config";
-import { probeProviderHealth } from "../../inference-health";
-import * as nim from "../../nim";
+import { parseGatewayInference } from "../../inference/config";
+import { probeProviderHealth } from "../../inference/health";
+import * as nim from "../../inference/nim";
 import * as onboardSession from "../../onboard-session";
 import type { Session } from "../../onboard-session";
 import {
