+++ src/lib/inference/model-prompts.ts
@@ -1,12 +1,12 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { CLOUD_MODEL_OPTIONS } from "./inference-config";
-import { isSafeModelId } from "./validation";
+import { CLOUD_MODEL_OPTIONS } from "./config";
+import { isSafeModelId } from "../validation";
 import { validateNvidiaEndpointModel } from "./provider-models";
 
 // credentials.ts still uses CommonJS-style exports.
-const { getCredential, prompt } = require("./credentials/store");
+const { getCredential, prompt } = require("../credentials/store");
 
 export const BACK_TO_SELECTION = "__NEMOCLAW_BACK_TO_SELECTION__";
