+++ src/lib/inference/nim.ts
@@ -4,7 +4,7 @@
 // NIM container management — pull, start, stop, health-check NIM images.
 
 const fs = require("fs");
-const { runCapture } = require("./runner");
+const { runCapture } = require("../runner");
 const {
   dockerContainerInspectFormat,
   dockerForceRm,
@@ -14,11 +14,11 @@ const {
   dockerRm,
   dockerRunDetached,
   dockerStop,
-} = require("./adapters/docker");
-const { sleepSeconds } = require("./core/wait");
-const nimImages = require("../../bin/lib/nim-images.json");
+} = require("../adapters/docker");
+const { sleepSeconds } = require("../core/wait");
+const nimImages = require("../../../bin/lib/nim-images.json");
 
-import { VLLM_PORT } from "./core/ports";
+import { VLLM_PORT } from "../core/ports";
 
 const UNIFIED_MEMORY_GPU_TAGS = ["GB10", "Thor", "Orin", "Xavier"];
 const NIM_STATUS_PROBE_TIMEOUT_MS = 5000;
