+++ src/lib/inference/ollama/windows.ts
@@ -6,9 +6,9 @@
 // Detection lives in onboard.ts; this module owns the action side.
 
 const { spawn, spawnSync } = require("child_process");
-const { run, runCapture } = require("./runner");
-const { OLLAMA_HOST_DOCKER_INTERNAL, setResolvedOllamaHost } = require("./local-inference");
-const { OLLAMA_PORT } = require("./core/ports");
+const { run, runCapture } = require("../../runner");
+const { OLLAMA_HOST_DOCKER_INTERNAL, setResolvedOllamaHost } = require("../local");
+const { OLLAMA_PORT } = require("../../core/ports");
 
 function sleep(seconds: number): void {
   spawnSync("sleep", [String(seconds)]);
