+++ src/lib/onboard-providers.ts
@@ -9,7 +9,7 @@ const {
   DEFAULT_CLOUD_MODEL,
   OLLAMA_LOCAL_CREDENTIAL_ENV,
   VLLM_LOCAL_CREDENTIAL_ENV,
-} = require("./inference-config");
+} = require("./inference/config");
 const { isSafeModelId } = require("./validation");
 const { compactText } = require("./core/url-utils");
