+++ test/ollama-tools-capability.test.ts
@@ -7,8 +7,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 const require = createRequire(import.meta.url);
 const REPO_ROOT = path.join(import.meta.dirname, "..");
-const LOCAL_INFERENCE_PATH = path.join(REPO_ROOT, "dist", "lib", "local-inference.js");
-const ONBOARD_OLLAMA_PROXY_PATH = path.join(REPO_ROOT, "dist", "lib", "onboard-ollama-proxy.js");
+const LOCAL_INFERENCE_PATH = path.join(REPO_ROOT, "dist", "lib", "inference", "local.js");
+const ONBOARD_OLLAMA_PROXY_PATH = path.join(REPO_ROOT, "dist", "lib", "inference", "ollama", "proxy.js");
 
 type CapturedCall = { argv: readonly string[]; opts?: Record<string, unknown> };
